feat: se agrego el jasper report para los PDF

+++ src/main/java/org/inventory_system/interfaces/SalesDAO.java
@@ -12,7 +12,7 @@ public interface SalesDAO {
     int calculateFinalAmount(int salId) throws SQLException;
     void deleteSale(int saleId);
     void deleteSalesDetails(int prodId, int saleId);
-    double getTotalSales();
+    int getTotalSales();
     double getSalesActualMonth(String month);
     int getSalesId();
 }
